requirements.txt
argparse
typing
transformers
pillow
tf-keras
protobuf
tensorflow
chardet
numpy
opencv-python
torch
